+++ NEWS.md
@@ -9,6 +9,11 @@ function. This allows for example to use
 `text = tags$span(style = "color:lime;", "Message")` instead of 
 `HTML('<span style="color:lime;">Message</span>')`.
 
+* The `showToast` function has a new argument `id`. See the details in 
+`?showToast` for its usage. It can be used for an alternative of the R callback.
+
+* New function `toastUpdate`, to update a toast.
+
 
 # shinyToastify 1.0.0